requirements.txt
aiogram>=3.3,<4.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
python-dotenv>=1.0,<2.0
supabase>=2.3,<3.0
